feat : 전체 공고 부분 적응형 적용 및 컴포넌트 분리
